The root tickets symlink is hidden from git the moment it is made
An uncommitted symlink at the repo root rides any sweeping git add -A onto
a code branch — the leak the drain E2E surfaced. The info/exclude mechanism
moves out of worktree-deps into git-exclude.ts so both callers share it;
the rule is root-anchored and only written when the link is ours, so a
user's own tickets path stays visible. The sibling leak (auto-onboard
skipping installProject) is #1600.
