+++ packages/the-framework/src/data-branch.test.SPEC.md
@@ -1,4 +1,4 @@
-Covers the data branch against real git: birth as a parentless history (or adoption of origin's), the seeded queue and the root tickets symlink (a user's own path left alone), funneled writes committing on the branch and pushing while main stays untouched, stranded local commits carried out and conflicting ones resolved toward origin with the op re-applied, the eager pull converging on what other machines pushed, and concurrent writes serializing whole.
+Covers the data branch against real git: birth as a parentless history (or adoption of origin's), the seeded queue and the root tickets symlink (hidden from git, while a user's own path is left alone and left visible), funneled writes committing on the branch and pushing while main stays untouched, stranded local commits carried out and conflicting ones resolved toward origin with the op re-applied, the eager pull converging on what other machines pushed, and concurrent writes serializing whole.
 
 ## Before modifying/creating SPEC.md files
